chore: full repository audit and contributor readiness overhaul
Repository audit and polish to prepare for external contributors.

Removed:
- contributor-candidates.md and outreach-emails/ (internal recruitment artifacts)
- gnn/ intermediate training checkpoints, logs, confusion matrix files
- gnn/__pycache__ directories

Added:
- LICENSE (MIT)
- .editorconfig (consistent formatting across Rust/Swift/Python/Markdown)
- CONTRIBUTING.md (full contributor guide with conventions)
- CODE_OF_CONDUCT.md (Contributor Covenant 2.1)
- SECURITY.md (vulnerability reporting policy + design principles)
- CHANGELOG.md (Keep a Changelog format, v1.0 → v2.1)
- ROADMAP.md (near/medium/long term plans)
- DEVELOPMENT.md (10-minute setup guide)
- GOOD_FIRST_ISSUES.md (beginner/medium/advanced task lists)
- docs/ARCHITECTURE.md (system design with Mermaid diagrams)
- docs/DESIGN_PRINCIPLES.md (10 core principles)
- docs/STYLE_GUIDE.md (Rust/Swift/Python/Markdown conventions)
- docs/README.md (documentation index)
- gnn/README.md (GNN architecture and training guide)
- examples/README.md + examples/configs/ (4 sample configs: default, gaming, development, conservative)
- examples/cli/README.md (CLI usage examples and shell one-liners)
- scripts/check.sh (run all pre-commit checks)
- scripts/build.sh (build full .app bundle)
- scripts/install.sh (install CLI binary)
- .github/workflows/ci.yml (CI: fmt, clippy, test, build on macOS + Linux, cross-compile check, Swift build)
- .github/ISSUE_TEMPLATE/ (bug report, feature request, question templates)
- .github/PULL_REQUEST_TEMPLATE.md

Modified:
- Cargo.toml: version 2.1.0, correct repo URL, keywords, categories
- rust-toolchain.toml: added Linux targets (x86_64 + aarch64)
- build.rs: deployment target 13.0, cfg(target_os = "macos") guard
- README.md: complete rewrite with accurate features, architecture, structure
- AGENTS.md: converted from agent-specific to contributor-facing guide
- .gitignore: ignore bundled CoreML model in Swift sources
- gui/build_app.sh: version bump to match Cargo.toml

Code quality:
- Fixed all clippy warnings (34 → 0) via auto-fix + manual cleanup
- Fixed all compiler warnings on macOS and Linux
- &PathBuf → &Path where clippy recommended
- sort_by → sort_by_key with Reverse
- Manual Iterator::find replaced with .find()
- Severity::from_str renamed to parse_severity (avoid trait confusion)
- #[allow(dead_code)] on public API methods used only by binary
- cargo fmt applied

Verified:
- cargo build: 0 warnings, 0 errors
- cargo clippy: 0 warnings
- cargo test: 327 tests pass
- cargo check --target x86_64-unknown-linux-gnu: 0 warnings
- cargo fmt --check: passes

+++ src/engines/disk/engine.rs
@@ -67,7 +67,7 @@ fn apfs_stats() -> Option<ApfsStats> {
 
     // "CapacityInUse" is the actual bytes used by this APFS volume.
     let system_used = parse_bytes(&root_str, "CapacityInUse").unwrap_or(0);
-    let data_used   = parse_bytes(&data_str, "CapacityInUse").unwrap_or(0);
+    let data_used = parse_bytes(&data_str, "CapacityInUse").unwrap_or(0);
 
     // Best-effort: sum physical sizes of /Applications and /System/Applications.
     let applications_used = dir_size_fast(std::path::Path::new("/Applications"))
@@ -96,7 +96,7 @@ fn linux_stats() -> Option<ApfsStats> {
         }
         let buf = buf.assume_init();
 
-        let block_size = buf.f_frsize as u64;
+        let block_size = buf.f_frsize;
         let total = buf.f_blocks as u64 * block_size;
         let free = buf.f_bavail as u64 * block_size;
         let used = (buf.f_blocks as u64 - buf.f_bfree as u64) * block_size;
@@ -215,22 +215,30 @@ impl Engine for DiskEngine {
         // without needing a separate API call.
         if let Some(stats) = volume_stats() {
             let total_known_used = stats.system_used + stats.data_used;
-            let vol_label = if cfg!(target_os = "macos") { "Macintosh HD" } else { "Root filesystem (/)" };
+            let vol_label = if cfg!(target_os = "macos") {
+                "Macintosh HD"
+            } else {
+                "Root filesystem (/)"
+            };
             let vol_finding = Finding::new(
                 EngineId::All,
                 Severity::Info,
                 Category::SystemInfo,
                 Target::Path(std::path::PathBuf::from("/")),
-                format!("Volume: {} total, {} used, {} free",
+                format!(
+                    "Volume: {} total, {} used, {} free",
                     crate::util::disk::format_bytes(stats.total),
                     crate::util::disk::format_bytes(total_known_used),
-                    crate::util::disk::format_bytes(stats.free)),
-                format!("{} — {} total capacity, {} data + {} system used, {} available",
+                    crate::util::disk::format_bytes(stats.free)
+                ),
+                format!(
+                    "{} — {} total capacity, {} data + {} system used, {} available",
                     vol_label,
                     crate::util::disk::format_bytes(stats.total),
                     crate::util::disk::format_bytes(stats.data_used),
                     crate::util::disk::format_bytes(stats.system_used),
-                    crate::util::disk::format_bytes(stats.free)),
+                    crate::util::disk::format_bytes(stats.free)
+                ),
             )
             .with_size(total_known_used)
             // Rich JSON so Swift can build accurate donut segments:
@@ -290,14 +298,23 @@ impl Engine for DiskEngine {
             entries.truncate(self.args.top);
 
             for (path, size, is_dir) in entries {
-                let name = path.file_name()
+                let name = path
+                    .file_name()
                     .map(|n| n.to_string_lossy().to_string())
                     .unwrap_or_else(|| path.to_string_lossy().to_string());
 
                 let title = if is_dir {
-                    format!("Disk usage: {} ({} dir)", name, crate::util::disk::format_bytes(size))
+                    format!(
+                        "Disk usage: {} ({} dir)",
+                        name,
+                        crate::util::disk::format_bytes(size)
+                    )
                 } else {
-                    format!("Disk usage: {} ({} file)", name, crate::util::disk::format_bytes(size))
+                    format!(
+                        "Disk usage: {} ({} file)",
+                        name,
+                        crate::util::disk::format_bytes(size)
+                    )
                 };
 
                 // SystemInfo category → not counted in reclaimable totals.
@@ -307,7 +324,13 @@ impl Engine for DiskEngine {
                     Category::SystemInfo,
                     Target::Path(path.clone()),
                     title,
-                    format!("{} '{}' in {} — {} (physical disk usage)", if is_dir { "Directory" } else { "File" }, name, search_path.display(), crate::util::disk::format_bytes(size)),
+                    format!(
+                        "{} '{}' in {} — {} (physical disk usage)",
+                        if is_dir { "Directory" } else { "File" },
+                        name,
+                        search_path.display(),
+                        crate::util::disk::format_bytes(size)
+                    ),
                 )
                 .with_size(size)
                 .with_hint(if is_dir {
@@ -343,7 +366,8 @@ impl Engine for DiskEngine {
             large_files.truncate(self.args.top);
 
             for (path, size) in large_files {
-                let name = path.file_name()
+                let name = path
+                    .file_name()
                     .map(|n| n.to_string_lossy().to_string())
                     .unwrap_or_else(|| path.to_string_lossy().to_string());
 
@@ -352,8 +376,16 @@ impl Engine for DiskEngine {
                     Severity::Info,
                     Category::LargeFile,
                     Target::Path(path.clone()),
-                    format!("Large file: {} ({})", name, crate::util::disk::format_bytes(size)),
-                    format!("File '{}' — {} (physical disk usage)", path.display(), crate::util::disk::format_bytes(size)),
+                    format!(
+                        "Large file: {} ({})",
+                        name,
+                        crate::util::disk::format_bytes(size)
+                    ),
+                    format!(
+                        "File '{}' — {} (physical disk usage)",
+                        path.display(),
+                        crate::util::disk::format_bytes(size)
+                    ),
                 )
                 .with_size(size)
                 .with_hint("Review and delete if no longer needed".to_string());
